requirements.txt
Django==3.2.8
pymysql==1.0.2
django-cors-headers==3.8.0
pyjwt==2.3.0
bcrypt==3.2.0